backend/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib
bcrypt==3.2.0
pydantic-settings
python-multipart
email-validator
